+++ src/App.js
@@ -11,7 +11,7 @@ import Reso from './Components/Reso';
 import ResoConfirmation from './Components/ResoConfirmation';
 import Favorite from './Components/Favorite';
 import Reservation from './Components/Reservation';
-import Account from './Components/Account';
+import Update from './Components/Update';
 import Calendar from './Components/Calendar';
 // import { SearchProvider } from './Components/searchContext';
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
@@ -138,8 +138,8 @@ function App() {
           <Favorite {...user} setUser={setUser}/>
         </Route> 
 
-        <Route path='/account'>
-          <Account {...user} setUser={setUser}/>
+        <Route path='/update'>
+          <Update {...user} setUser={setUser}/>
         </Route>
 
         {/* for coding purposes */}
