--- src/Components/Account.jsx
@@ -1,11 +1,0 @@
-import React, { useState, useEffect, useRef } from 'react';
-
-function Account({user, setUser}) {
-    return(
-        <div>
-            <h2 className='fancy'>hello</h2>
-        </div>
-    )
-}
-
-export default Account
