+++ src/css/forms.css
@@ -22,7 +22,4 @@
     left: 42.5%; */
     margin-left: 1rem;
 }
-.prl {
-    margin-top: 50px;
-    padding: 1rem;
-}
+
